Address Gemini follow-up: ss id cache, int32 scalars, plif guard

+++ spikingjelly/activation_based/cuda_kernel/auto_cuda/ss_neuron_kernel/ss_neuron_kernel_base.py
@@ -45,7 +45,7 @@ def scalar_to_cupy(py_dict: dict, ref: str = "x"):
                 py_dict[key] = value
 
             elif isinstance(value, int):
-                py_dict[key] = cupy.asarray(value)
+                py_dict[key] = cupy.asarray(value, dtype=np.int32)
 
 
 def new_tensors(news: tuple, py_dict: dict, ref: str = "x"):
